feat: add hook for parsing serach param
